+++ yt_dlp/downloader/fragment.py
@@ -12,8 +12,8 @@ except ImportError:
 
 from .common import FileDownloader
 from .http import HttpFD
+from ..aes import compat_aes_cbc_decrypt
 from ..compat import (
-    compat_aes_cbc_decrypt,
     compat_urllib_error,
     compat_struct_pack,
 )
